+++ src/app/features/wallets/wallets.page.ts
@@ -8,7 +8,7 @@ import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
 import { NgxQrcodeElementTypes } from '@techiediaries/ngx-qrcode';
 import { BehaviorSubject, fromEvent } from 'rxjs';
 import { concatMap, first, map, tap } from 'rxjs/operators';
-import { WebCryptoApiSignatureProvider } from '../../shared/collector/signature/web-crypto-api-signature-provider/web-crypto-api-signature-provider.service';
+import { CaptureAppWebCryptoApiSignatureProvider } from '../../shared/collector/signature/capture-app-web-crypto-api-signature-provider/capture-app-web-crypto-api-signature-provider.service';
 import { DiaBackendAuthService } from '../../shared/dia-backend/auth/dia-backend-auth.service';
 import { BUBBLE_IFRAME_URL } from '../../shared/dia-backend/secret';
 import { DiaBackendWalletService } from '../../shared/dia-backend/wallet/dia-backend-wallet.service';
@@ -20,8 +20,8 @@ import { BubbleToIonicPostMessage } from '../../shared/iframe/iframe';
   styleUrls: ['./wallets.page.scss'],
 })
 export class WalletsPage {
-  readonly publicKey$ = this.webCryptoApiSignatureProvider.publicKey$;
-  readonly privateKey$ = this.webCryptoApiSignatureProvider.privateKey$;
+  readonly publicKey$ = this.capAppWebCryptoApiSignatureProvider.publicKey$;
+  readonly privateKey$ = this.capAppWebCryptoApiSignatureProvider.privateKey$;
   readonly assetWalletAddr$ = this.diaBackendWalletService.assetWalletAddr$;
 
   readonly networkConnected$ = this.diaBackendWalletService.networkConnected$;
@@ -43,12 +43,10 @@ export class WalletsPage {
     private readonly diaBackendAuthService: DiaBackendAuthService,
     private readonly snackBar: MatSnackBar,
     private readonly translocoService: TranslocoService,
-    private readonly webCryptoApiSignatureProvider: WebCryptoApiSignatureProvider,
+    private readonly capAppWebCryptoApiSignatureProvider: CaptureAppWebCryptoApiSignatureProvider,
     private readonly router: Router,
     private readonly navController: NavController
-  ) {}
-
-  ionViewDidEnter() {
+  ) {
     this.processIframeEvents();
   }
 
@@ -63,7 +61,7 @@ export class WalletsPage {
               this.iframeLoaded$.next(true);
               break;
             case BubbleToIonicPostMessage.IFRAME_BACK_BUTTON_CLICKED:
-              this.navController.pop();
+              this.navController.back();
               break;
             case BubbleToIonicPostMessage.IFRAME_BUY_NUM_BUTTON_CLICKED:
               this.navigateToBuyNumPage();