+++ src/app/shared/collector/signature/capture-app-web-crypto-api-signature-provider/capture-app-web-crypto-api-signature-provider.service.spec.ts
@@ -2,18 +2,22 @@ import { TestBed } from '@angular/core/testing';
 import { defer } from 'rxjs';
 import { concatMapTo } from 'rxjs/operators';
 import { sortObjectDeeplyByKey } from '../../../../utils/immutable/immutable';
-import { isSignature, SignedMessage } from '../../../repositories/proof/proof';
+import {
+  isSignature,
+  RecorderType,
+  SignedMessage,
+} from '../../../repositories/proof/proof';
 import { SharedTestingModule } from '../../../shared-testing.module';
-import { WebCryptoApiSignatureProvider } from './web-crypto-api-signature-provider.service';
+import { CaptureAppWebCryptoApiSignatureProvider } from './capture-app-web-crypto-api-signature-provider.service';
 
-describe('WebCryptoApiSignatureProvider', () => {
-  let provider: WebCryptoApiSignatureProvider;
+describe('CaptureAppWebCryptoApiSignatureProvider', () => {
+  let provider: CaptureAppWebCryptoApiSignatureProvider;
 
   beforeEach(() => {
     TestBed.configureTestingModule({
       imports: [SharedTestingModule],
     });
-    provider = TestBed.inject(WebCryptoApiSignatureProvider);
+    provider = TestBed.inject(CaptureAppWebCryptoApiSignatureProvider);
   });
 
   it('should be created', () => expect(provider).toBeTruthy());
@@ -58,7 +62,7 @@ describe('WebCryptoApiSignatureProvider', () => {
   it('should provide signature', async () => {
     const signedMessage: SignedMessage = {
       spec_version: '',
-      recorder: '',
+      recorder: RecorderType.Capture,
       created_at: 0,
       proof_hash: '',
       asset_mime_type: '',