+++ src/app/shared/migration/service/migration.service.ts
@@ -1,6 +1,7 @@
 import { HttpErrorResponse } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
+import { CameraSource } from '@capacitor/camera';
 import { defer, forkJoin, iif } from 'rxjs';
 import {
   catchError,
@@ -12,7 +13,7 @@ import {
 } from 'rxjs/operators';
 import { VOID$ } from '../../../utils/rx-operators/rx-operators';
 import { CollectorService } from '../../collector/collector.service';
-import { WebCryptoApiSignatureProvider } from '../../collector/signature/web-crypto-api-signature-provider/web-crypto-api-signature-provider.service';
+import { CaptureAppWebCryptoApiSignatureProvider } from '../../collector/signature/capture-app-web-crypto-api-signature-provider/capture-app-web-crypto-api-signature-provider.service';
 import {
   DiaBackendAsset,
   DiaBackendAssetRepository,
@@ -46,7 +47,7 @@ export class MigrationService {
     private readonly preferenceManager: PreferenceManager,
     private readonly onboardingService: OnboardingService,
     private readonly versionService: VersionService,
-    private readonly webCryptoApiSignatureProvider: WebCryptoApiSignatureProvider
+    private readonly capAppWebCryptoApiSignatureProvider: CaptureAppWebCryptoApiSignatureProvider
   ) {}
 
   migrate$(skip?: boolean) {
@@ -161,7 +162,7 @@ export class MigrationService {
           err.status === HttpErrorCode.NOT_FOUND
         ) {
           return defer(() =>
-            this.webCryptoApiSignatureProvider.getPrivateKey()
+            this.capAppWebCryptoApiSignatureProvider.getPrivateKey()
           ).pipe(
             concatMap(privateKey =>
               this.diaBackendWalletService.setIntegrityWallet$(privateKey)
@@ -171,7 +172,7 @@ export class MigrationService {
         throw err;
       }),
       concatMap(assetWallet =>
-        this.webCryptoApiSignatureProvider.importKeys(
+        this.capAppWebCryptoApiSignatureProvider.importKeys(
           assetWallet.address,
           assetWallet.private_key
         )
@@ -185,7 +186,9 @@ export class MigrationService {
       map(proofs => proofs.filter(proof => !proof.signatureVersion)),
       concatMap(proofs =>
         forkJoin(
-          proofs.map(proof => this.collectorService.generateSignature(proof))
+          proofs.map(proof =>
+            this.collectorService.generateSignature(proof, CameraSource.Camera)
+          )
         ).pipe(defaultIfEmpty(proofs))
       ),
       concatMap(proofs =>